feat: add consumer
